extract getContrastColor to colorUtils. refactor calendar components to reflect change.
