+++ src/fetching/getCampaign.ts
@@ -1,5 +1,5 @@
 import { defaultFetch } from '.';
-import { generateRandomColor } from '../utils/generateRandomColor';
+import { generateRandomColor } from '../utils/colorUtils';
 import { ZetkinCampaign } from '../types/zetkin';
